feat(auth): add support for passport and authentication process purposal
This commit add the ability to have built-in authentication process in a
hand clap time.
It relies on passport, jsonwebtoken, and mongo-connect.

It implements several features, for example an Auth service on client
side with ready to use methods such `login` and `signup`.

On a purposed future, this authentication process could be enrich with
some Oauth methods (connect with google, github...) or maybe a simple
lognup? To be continued...